# swagger.yaml
openapi: "3.0.0"
info:
  title: "Portfolio API"
  version: "1.0.0"
  description: "API for managing portfolio projects and timeline entries"
servers:
  url: "http://localhost:3000"
  description: "Local server"
apis: ["./routes/*.js"]
paths:
  /api:
    get:
      summary: Retrieve a description of the API
      responses:
        200:
          description: A descriptive JSON of the API
          content:
            application/json:
              schema:
                type: JSON
  /projects:
    post:
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./models/project.model.js/ProjectSchema"
      responses:
        201:
          description: The created project
          content:
            application/json:
              schema:
                $ref: "./models/project.model.js/ProjectSchema"
  /projects:
    get:
      summary: Retrieve a list of projects
      responses:
        200:
          description: A list of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "./models/project.model.js/ProjectSchema"
  /project/{id}:
    get:
      summary: Retrieve a single project by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The project ID
      responses:
        200:
          description: A single project
          content:
            application/json:
              schema:
                $ref: "./models/project.model.js/ProjectSchema"
  /project/{id}:
    put:
      summary: Update an existing project by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The project ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./models/project.model.js/ProjectSchema"
      responses:
        201:
          description: The updated project
          content:
            application/json:
              schema:
                $ref: "./models/project.model.js/ProjectSchema"

  /project/{id}:
    delete:
      summary: Delete a project by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The project ID
      responses:
        200:
          description: Project deleted
          content:
            application/json:
              schema:
                type: JSON